switch to muxrpc-validation 2.0
