+++ lib/cli-cmd-aliases.js
