+++ package.json
@@ -24,7 +24,7 @@
     "multiblob": "^1.5.0",
     "multicb": "^1.0.0",
     "muxrpc": "^6.1.1",
-    "muxrpc-validation": "^1.0.0",
+    "muxrpc-validation": "^2.0.0",
     "muxrpcli": "^1.0.0",
     "mynosql-query": "~1.0.0",
     "nomnom": "1.8.0",
